fix(assignment-assessor): RubricLevel.points int → float
Genkit/TS fixtures send fractional points (1.5, 0.5) but the Pydantic
schema declared int, causing every assess request to fail with
`int_from_float` validation errors. Mirrors the float type on
PerCriterionScore.points + maxPoints.

Also deselect tests/unit/test_action_flow_ts_parity.py from Cloud
Build — the test reads sahayakai-main/src/lib/sidecar/types.generated.ts
which isn't mounted in the build container. Pre-existing CI hole;
keeping the rest of tests/unit running.

+++ sahayakai-agents/src/sahayakai_agents/agents/assignment_assessor/schemas.py
@@ -27,7 +27,7 @@ class RubricLevel(BaseModel):
 
     name: str
     description: str
-    points: int = Field(ge=0)
+    points: float = Field(ge=0)
 
 
 class RubricCriterion(BaseModel):
